requirements.txt
Django==1.7.8
django-autoslug==1.8.0
django-debug-toolbar==1.3.2
django-extensions==1.5.5
django-grappelli==2.6.5
django-imagekit==3.2.6
Pillow==2.9.0
